requirements.txt
APScheduler==3.10.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-auth==2.16.2
google.cloud
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rsa==4.9
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
